Add and document ArgoIndex
