+++ argopy/__init__.py
@@ -33,7 +33,7 @@ from .utilities import TopoFetcher, ArgoNVSReferenceTables, OceanOPSDeployments 
 from .utilities import monitor_status as status  # noqa: E402
 from .options import set_options  # noqa: E402
 from .data_fetchers import CTDRefDataFetcher
-
+from .stores import ArgoIndex
 
 #
 __all__ = (
@@ -54,7 +54,8 @@ __all__ = (
     "TopoFetcher",  # Class
     "ArgoNVSReferenceTables",  # Class
     "OceanOPSDeployments",  # Class
-    "CTDRefDataFetcher",
+    "CTDRefDataFetcher",  # Class
+    "ArgoIndex",  # Class
 
     # Submodules:
     "utilities",
