+++ argopy/stores/__init__.py
@@ -1,16 +1,14 @@
-from .argo_index import indexstore, indexfilter_wmo, indexfilter_box
 from .filesystems import filestore, httpstore, memorystore, ftpstore, httpstore_erddap, httpstore_erddap_auth
 
 from .argo_index_pa import indexstore_pyarrow as indexstore_pa
 from .argo_index_pd import indexstore_pandas as indexstore_pd
+from .argo_index import ArgoIndex
 
 
 #
 __all__ = (
     # Classes:
-    "indexstore",
-    "indexfilter_wmo",
-    "indexfilter_box",
+    "ArgoIndex",
     "indexstore_pa",
     "indexstore_pd",
     "filestore",
