+++ argopy/stores/argo_index_pd.py
@@ -23,7 +23,7 @@ class indexstore_pandas(ArgoIndexStoreProto):
     With this store, index and search results are saved as pickle files in cache
 
     """
-    __doc__ += ArgoIndexStoreProto.__doc__
+    # __doc__ += ArgoIndexStoreProto.__doc__
 
     backend = "pandas"
