+++ argopy/stores/argo_index_proto.py
@@ -2,8 +2,6 @@
 Argo file index store prototype
 
 """
-
-import os
 import numpy as np
 import pandas as pd
 import logging
@@ -29,56 +27,6 @@ log = logging.getLogger("argopy.stores.index")
 
 
 class ArgoIndexStoreProto(ABC):
-    """
-
-        Examples
-        --------
-
-        An index store is instantiated with the access path (host) and the index file:
-
-        >>> idx = indexstore()
-        >>> idx = indexstore(host="ftp://ftp.ifremer.fr/ifremer/argo")
-        >>> idx = indexstore(host="https://data-argo.ifremer.fr", index_file="ar_index_global_prof.txt")
-        >>> idx = indexstore(host="https://data-argo.ifremer.fr", index_file="ar_index_global_prof.txt", cache=True)
-
-        Full index methods and properties:
-
-        >>> idx.load()
-        >>> idx.load(nrows=12)  # Only load the first N rows of the index
-        >>> idx.N_RECORDS  # Shortcut for length of 1st dimension of the index array
-        >>> idx.index  # internal storage structure of the full index (:class:`pyarrow.Table` or :class:`pandas.DataFrame`)
-        >>> idx.shape  # shape of the full index array
-        >>> idx.convention  # What is the expected index format (core vs BGC profile index)
-        >>> idx.uri_full_index  # List of absolute path to files from the full index table column 'file'
-        >>> idx.to_dataframe(index=True)  # Convert index to user-friendly :class:`pandas.DataFrame`
-        >>> idx.to_dataframe(index=True, nrows=2)  # Only returns the first nrows of the index
-
-        Search methods and properties:
-
-        >>> idx.search_wmo(1901393)
-        >>> idx.search_cyc(1)
-        >>> idx.search_wmo_cyc(1901393, [1,12])
-        >>> idx.search_tim([-60, -55, 40., 45., '2007-08-01', '2007-09-01'])  # Take an index BOX definition
-        >>> idx.search_lat_lon([-60, -55, 40., 45., '2007-08-01', '2007-09-01'])  # Take an index BOX definition
-        >>> idx.search_lat_lon_tim([-60, -55, 40., 45., '2007-08-01', '2007-09-01'])  # Take an index BOX definition
-        >>> idx.search_params(['C1PHASE_DOXY', 'DOWNWELLING_PAR'])  # Take a list of strings, only for BGC index !
-        >>> idx.N_MATCH  # Shortcut for length of 1st dimension of the search results array
-        >>> idx.search  # Internal table with search results
-        >>> idx.uri  # List of absolute path to files from the search results table column 'file'
-        >>> idx.run()  # Run the search and save results in cache if necessary
-        >>> idx.to_dataframe()  # Convert search results to user-friendly :class:`pandas.DataFrame`
-        >>> idx.to_dataframe(nrows=2)  # Only returns the first nrows of the search results
-        >>> idx.to_indexfile("search_index.txt")  # Export search results to Argo standard index file
-
-        Misc:
-
-        >>> idx.cname
-        >>> idx.read_wmo
-        >>> idx.read_params
-        >>> idx.records_per_wmo
-
-    """
-
     backend = "?"
     """Name of store backend"""  # Pandas or Parquet
